+++ piGantry/piSerial/mathFunc.py
@@ -1,2 +1,3 @@
-class inputProcess(object):
-    
+class inputProcess():
+    def __init__(self):
+        self.steps